requirements.txt
python-dotenv
eth_account